auth 패키지 정리
기존에 있던 인증 관련된 기능을 auth 패키지에 정리하였습니다.
