+++ src/main/java/com/appsolve/wearther_backend/Controller/MemberController.java
@@ -5,6 +5,7 @@ import com.appsolve.wearther_backend.apiResponse.ApiResponse;
 import com.appsolve.wearther_backend.Dto.SignInRequest;
 
 import com.appsolve.wearther_backend.Service.MemberService;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -22,19 +23,14 @@ import java.util.Map;
 public class MemberController {
     private final MemberService memberService;
 
+
     @PostMapping("/signUp")
     public ResponseEntity<?> signUp(@Valid @RequestBody SignUpRequest request) {
-         Long memberId = memberService.registerMember(request);
+        Long memberId = memberService.registerMember(request);
         return ApiResponse.success(HttpStatus.
                 CREATED,Map.of("memberId", memberId));
     }
 
-    @PostMapping("/login")
-    public ResponseEntity<?> login(@RequestBody SignInRequest request) {
-        Map<String, ?> authMap = memberService.signInMember(request);
-        String token =(String) authMap.remove("accessToken");
-        return ApiResponse.loginSuccess(HttpStatus.OK, authMap,token);
-    }
 
     @GetMapping("/duplication-check")
     public  ResponseEntity<?> checkDuplication(@RequestParam String loginId) {
