+++ src/main/java/com/appsolve/wearther_backend/auth/Dto/ReissueDto.java
@@ -1,0 +1,9 @@
+package com.appsolve.wearther_backend.auth.Dto;
+
+import lombok.Data;
+
+@Data
+public class ReissueDto {
+    private String refreshToken;
+    private String deviceId;
+}
