+++ src/main/java/com/appsolve/wearther_backend/closet/service/ClosetService.java
@@ -2,8 +2,7 @@ package com.appsolve.wearther_backend.closet.service;
 
 import com.appsolve.wearther_backend.Entity.MemberEntity;
 import com.appsolve.wearther_backend.Repository.MemberRepository;
-import com.appsolve.wearther_backend.Service.AuthService;
-import com.appsolve.wearther_backend.Service.MemberService;
+import com.appsolve.wearther_backend.auth.Service.AuthService;
 import com.appsolve.wearther_backend.Service.TasteService;
 import com.appsolve.wearther_backend.closet.ShoppingUrls;
 import com.appsolve.wearther_backend.closet.dto.ClosetResponseDto;
